requirements.txt
pandas
pyspark
kafka-python
psycopg2-binary
